Mal/generate test data

See merge request adoptapal/adoptapalwebapplication!12

+++ Adoptapal/Adoptapal.Web/Controllers/AccountController.cs
@@ -1,4 +1,14 @@
-﻿using Adoptapal.Business.Definitions;
+﻿/*
+ * File: AccountController.cs
+ * Namespace: Adoptapal.Web.Controllers
+ * 
+ * Description:
+ * This file contains the implementation of the AccountController class, which handles
+ * user account-related actions such as login, registration, password reset, and user settings.
+ * 
+ */
+
+using Adoptapal.Business.Definitions;
 using Adoptapal.Business.Implementations;
 using Microsoft.AspNetCore.Mvc;
 using System.Net.Mail;
@@ -244,16 +254,20 @@ namespace Adoptapal.Web.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteUser()
         {
-            User user = await _manager.GetUserByIdAsync(userId);
-
-            if(user != null)
+            if(userId != null)
             {
-                await _manager.DeleteUserAsync(user.Id);
+                User user = await _manager.GetUserByIdAsync(userId);
 
-                return RedirectToAction("Signout");
+                if (user != null)
+                {
+                    await _manager.DeleteUserAsync(user.Id);
+
+                    return RedirectToAction("Signout");
+                }
+
             }
-            return RedirectToAction("Login");
 
+            return RedirectToAction("Login");
         }
     }
 }
